requirements.txt
asgiref==3.7.2
bleach==6.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-cors-headers==4.1.0
django-rest-durin==1.1.0
djangorestframework==3.14.0
drf-api-logger==1.1.14
humanize==4.7.0
idna==3.4
pdfreader==0.1.12
psycopg2-binary==2.9.6
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.0
urllib3==2.0.3
webencodings==0.5.1
